Get rid of C-style type casting Corresponding PVS-Studio report:
Medium: Custom:
V2005: session.h:24 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: compatibility.cpp:50 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: compatibility.cpp:52 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: hexdump.cpp:16 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: restrictions.cpp:48 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: report.cpp:71 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: uconvert.cpp:80 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2004: uconvert.cpp:80 - Explicit conversion from 'long double' type to unsigned integer type.
V2005: uconvert.cpp:94 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: uconvert.cpp:100 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: uconvert.cpp:102 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: uconvert.cpp:188 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: uconvert.cpp:244 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2003: uconvert.cpp:244 - Explicit conversion from 'long double' type to signed integer type: (long) convert(from, to, value)
V2005: delegate.cpp:63 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: delegate.cpp:66 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: system_pipe.cpp:102 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: system_pipe.cpp:113 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: session.cpp:28 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: platform.cpp:66 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: platform.cpp:79 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: platform.cpp:85 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: platform.cpp:87 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: platform.cpp:102 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: platform.cpp:316 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: platform.cpp:317 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:45 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:47 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:65 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:67 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:89 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:159 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:205 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:335 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:348 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: runner.cpp:407 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: securerunner.cpp:92 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: securerunner.cpp:99 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: securerunner.cpp:130 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: securerunner.cpp:144 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_new.cpp:43 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_new.cpp:71 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_new.cpp:68 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_new.cpp:96 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2004: spawner_new.cpp:96 - Explicit conversion from 'double' type to unsigned integer type.
V2005: spawner_new.cpp:108 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_new.cpp:105 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_new.cpp:167 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_new.cpp:189 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_pcms2.cpp:39 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_pcms2.cpp:40 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_pcms2.cpp:46 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
V2005: spawner_pcms2.cpp:54 - C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.

+++ libspawner/inc/options.h
@@ -14,7 +14,7 @@
 
 struct options_class
 {
-    static const unsigned SHARED_MEMORY_BUF_SIZE = 4069;
+    static const size_t SHARED_MEMORY_BUF_SIZE = 4069;
 
     session_class session;
     std::string report_file;  /* bad, need many values
